+++ lib/bronze/commands/create_page/create_page_command.dart
@@ -1,11 +1,20 @@
-part of '../../amber.dart';
+part of '../../bronze.dart';
 
-void _createBottomSheetStructure() {
+void _createPageStructure() {
+  // Prepare route data
+  final prepared = _prepareRouteData();
+  if (!prepared) return;
+
+  final isAlreadyExists = _checkExistingRoute();
+  if (isAlreadyExists) {
+    print('❌ Error: Route already exists');
+    return;
+  }
   _moduleName = getModuleName();
   _role = '';
 
-  _toBeCreated = 'bs';
-  _pagePathType = 'bottom_sheets';
+  _toBeCreated = 'screen';
+  _pagePathType = 'screens';
 
   _contentPath = 'lib/$_pagePathType/$_pageName';
   _blocPath = '$_contentPath/bloc';
@@ -17,11 +26,6 @@ void _createBottomSheetStructure() {
   createDirectory(_blocPath);
   createDirectory(_viewPath);
   createDirectory(_componentsPath);
-
-  // Prepare route data
-  final prepared = _prepareRouteData();
-
-  if (!prepared) return;
 
   // import lines of view and bloc for appRoutes and route names
   _preparePageImports();
@@ -36,7 +40,7 @@ void _createBottomSheetStructure() {
   createFile('$_blocPath/${_pageName}_state.dart', _generateStateContent(_pageName));
 
   // Create View file
-  createFile('$_viewPath/${_pageName}_$_toBeCreated.dart', _generateViewContent(_pageName));
+  createFile('$_viewPath/${_pageName}_$_toBeCreated.dart', _generateScreenContent(_pageName));
 
   print('✅  $_pagePathType structure for "$_pageName" created successfully!');
 }