+++ lib/bronze/data/bronze_vars.dart
@@ -1,0 +1,24 @@
+part of '../bronze.dart';
+
+
+late String _moduleName;
+late String _role;
+late String _pageName; // name of the page/bottom sheet
+late String _variableName;
+
+late String _toBeCreated;
+late String _pagePathType;
+late String _contentPath;
+late String _blocPath;
+late String _viewPath;
+late String _componentsPath;
+
+late String _routesFilePath;
+late String _routeNamesFilePath;
+
+late File _routesFile;
+late File _routeNamesFile;
+
+// import lines for view and bloc
+late String _viewImport;
+late String _blocImport;