+++ lib/bronze/templates/route_name_template.dart
@@ -1,0 +1,6 @@
+part of '../bronze.dart';
+
+String _generateRouteNameEntry(String pageName) {
+  final String constantName = toCamelCase(pageName);
+  return "  static const String $constantName = '/$pageName';";
+}