+++ lib/bronze/templates/route_template.dart
@@ -1,0 +1,15 @@
+part of '../bronze.dart';
+
+String _generateRouteEntry(String pageName) {
+  final String className = toPascalCase(pageName);
+  final String constantName = toCamelCase(pageName);
+
+  return '''
+      AppRoute(
+        name: Routes.$constantName,
+        blocProvider: BlocPageProvider<${className}Bloc>(
+          bloc: (context) => ${className}Bloc(),
+          page: const ${className}Screen(),
+        ),
+      ),''';
+}