+++ lib/bronze/templates/view_template.dart
@@ -1,0 +1,56 @@
+part of '../bronze.dart';
+
+String _generateScreenContent(String pageName) {
+  String className = toPascalCase(pageName);
+  return '''
+import 'package:flutter/material.dart';
+$_blocImport
+import 'package:$_moduleName/widgets/common_appbar.dart';
+
+class ${className}Screen extends StatefulWidget {
+  const ${className}Screen({super.key});
+
+  @override
+  State<${className}Screen> createState() => _${className}State();
+}
+
+class _${className}State extends State<${className}Screen> {
+  late final bloc = context.${_variableName}Bloc;
+
+  @override
+  Widget build(BuildContext context) {
+    return Scaffold(
+      appBar: CommonAppbar(title: '$className Screen'),
+      body: Center(child: Text('Welcome to $className Screen!')),
+    );
+  }
+}
+''';
+}
+
+String _generateBottomSheetContent(String pageName) {
+  String className = toPascalCase(pageName);
+  return '''
+import 'package:flutter/material.dart';
+$_blocImport
+
+class ${className}BottomSheet extends StatefulWidget {
+  const ${className}BottomSheet({super.key});
+
+  @override
+  State<${className}BottomSheet> createState() => _${className}State();
+}
+
+class _${className}State extends State<${className}BottomSheet> {
+  late final bloc = context.${_variableName}Bloc;
+
+  @override
+  Widget build(BuildContext context) {
+    return const Padding(
+      padding: EdgeInsets.all(20),
+      child: Text("$className Bottom Sheet"),
+    );
+  }
+}
+''';
+}