+++ lib/bronze/utils/check_existing_route.dart
@@ -1,0 +1,7 @@
+part of '../bronze.dart';
+
+bool _checkExistingRoute() {
+  final String routeNamesContent = _routeNamesFile.readAsStringSync();
+  final String newRouteNameEntry = _generateRouteNameEntry(_pageName);
+  return routeNamesContent.contains(newRouteNameEntry);
+}