+++ lib/cli.dart
@@ -1,4 +1,5 @@
 import 'package:sg_cli/amber/amber.dart';
+import 'package:sg_cli/bronze/bronze.dart';
 import 'package:sg_cli/constants/constants.dart';
 import 'package:sg_cli/utils/config_helper.dart';
 
@@ -12,6 +13,8 @@ void runCLI(List<String> args) {
 
   if (config.version == kAmber) {
     Amber.runCommand();
+  } else if (config.version == kBronze) {
+    Bronze.runCommand();
   } else {
     print("❌  Error: Invalid config file sg_cli.yaml");
   }